done with the basic BookingFactory
